chore(refactor): enhanced general project structure (#88)
* update: store into stores folder


* update(editors): development


* update: development


* update: development


* update: development


* chore: utils >>> src


* chore: dev fixes


* fix: reslove issues


---------
