+++ package.json
@@ -22,11 +22,12 @@
     "@accordproject/markdown-transform": "^0.16.19",
     "@accordproject/template-engine": "^2.3.5",
     "@ant-design/icons": "^5.3.7",
-    "@monaco-editor/react": "^4.5.1",
+    "@monaco-editor/react": "^4.6.0",
     "antd": "^5.7.2",
     "core-js": "^3.37.1",
     "immer": "^10.0.2",
     "lz-string": "^1.5.0",
+    "monaco-editor": "^0.50.0",
     "node-stdlib-browser": "^1.2.0",
     "normalize.css": "^8.0.1",
     "react": "^18.2.0",
