+++ src/App.tsx
@@ -2,14 +2,14 @@ import { useEffect, useState } from "react";
 import { App as AntdApp, Layout, Row, Col, Collapse, theme, Grid } from "antd";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
-import AgreementData from "./AgreementData";
+import AgreementData from "./editors/editorsContainer/AgreementData";
 import AgreementHtml from "./AgreementHtml";
-import Errors from "./Errors";
-import TemplateMarkdown from "./TemplateMarkdown";
-import TemplateModel from "./TemplateModel";
-import useAppStore from "./store";
-import SampleDropdown from "./SampleDropdown";
-import FullScreenModal from "./FullScreenModal";
+import Errors from "./utils/helpers/Errors";
+import TemplateMarkdown from "./editors/editorsContainer/TemplateMarkdown";
+import TemplateModel from "./editors/editorsContainer/TemplateModel";
+import useAppStore from "./store/store";
+import SampleDropdown from "./components/SampleDropdown";
+import FullScreenModal from "./components/FullScreenModal";
 import { useSearchParams } from "react-router-dom";
 import UseShare from "./components/UseShare";
