+++ src/editors/editorsContainer/TemplateMarkdown.tsx
@@ -1,5 +1,5 @@
-import MarkdownEditor from "./MarkdownEditor";
-import useAppStore from "./store";
+import MarkdownEditor from "../MarkdownEditor";
+import useAppStore from "../../store/store";
 
 function TemplateMarkdown() {
   const templateMarkdown = useAppStore((state) => state.templateMarkdown);
